New design of "Team" subpage
The subpage containing infomration about the staff of the facility
was completely redesigned. The staff gets presented by their
kindergarten groups in a grid layout. Every group has its own
card which is turnable if you hover over the card. The front side
of the card presents a stock photo associated with the group, while
the user can read the names of the staff on the backside of the card.